+++ src/molpy/compute/pmsd.py
@@ -108,8 +108,8 @@ class PMSDCompute(Compute):
             elems_list.append(atoms["type"])
 
             # Get box
-            if frame.simbox is not None:
-                box_list.append(frame.simbox)
+            if frame.box is not None:
+                box_list.append(frame.box)
             else:
                 raise ValueError("Frame must contain box information")